SessionTest: split `request_multiple()` specific tests to dedicated test class
