+++ tests/Session/SessionTest.php
@@ -2,11 +2,8 @@
 
 namespace WpOrg\Requests\Tests\Session;
 
-use WpOrg\Requests\Exception\InvalidArgument;
-use WpOrg\Requests\Response;
 use WpOrg\Requests\Session;
 use WpOrg\Requests\Tests\TestCase;
-use WpOrg\Requests\Tests\TypeProviderHelper;
 
 final class SessionTest extends TestCase {
 	public function testURLResolution() {
@@ -127,37 +124,6 @@ final class SessionTest extends TestCase {
 		$this->assertSame('PATCH', $data['headers']['X-Requests-Request']);
 	}
 
-	public function testMultiple() {
-		$session   = new Session(httpbin('/'), ['X-Requests-Session' => 'Multiple']);
-		$requests  = [
-			'test1' => [
-				'url' => httpbin('/get'),
-			],
-			'test2' => [
-				'url' => httpbin('/get'),
-			],
-		];
-		$responses = $session->request_multiple($requests);
-
-		// test1
-		$this->assertNotEmpty($responses['test1']);
-		$this->assertInstanceOf(Response::class, $responses['test1']);
-		$this->assertSame(200, $responses['test1']->status_code);
-
-		$result = json_decode($responses['test1']->body, true);
-		$this->assertSame(httpbin('/get'), $result['url']);
-		$this->assertEmpty($result['args']);
-
-		// test2
-		$this->assertNotEmpty($responses['test2']);
-		$this->assertInstanceOf(Response::class, $responses['test2']);
-		$this->assertSame(200, $responses['test2']->status_code);
-
-		$result = json_decode($responses['test2']->body, true);
-		$this->assertSame(httpbin('/get'), $result['url']);
-		$this->assertEmpty($result['args']);
-	}
-
 	public function testSharedCookies() {
 		$session = new Session(httpbin('/'));
 
@@ -180,62 +146,5 @@ final class SessionTest extends TestCase {
 			'requests-testcookie' => 'testvalue',
 		];
 		$this->assertSame($cookies, $data['cookies']);
-	}
-
-	/**
-	 * Tests receiving an exception when the request_multiple() method received an invalid input type as `$requests`.
-	 *
-	 * @dataProvider dataRequestMultipleInvalidRequests
-	 *
-	 * @covers \WpOrg\Requests\Session::request_multiple
-	 *
-	 * @param mixed $input Invalid parameter input.
-	 *
-	 * @return void
-	 */
-	public function testRequestMultipleInvalidRequests($input) {
-		$this->expectException(InvalidArgument::class);
-		$this->expectExceptionMessage('Argument #1 ($requests) must be of type array|ArrayAccess&Traversable');
-
-		$session = new Session();
-		$session->request_multiple($input);
-	}
-
-	/**
-	 * Data Provider.
-	 *
-	 * @return array
-	 */
-	public function dataRequestMultipleInvalidRequests() {
-		$except = array_intersect(TypeProviderHelper::GROUP_ITERABLE, TypeProviderHelper::GROUP_ARRAY_ACCESSIBLE);
-		return TypeProviderHelper::getAllExcept($except);
-	}
-
-	/**
-	 * Tests receiving an exception when the request_multiple() method received an invalid input type as `$option`.
-	 *
-	 * @dataProvider dataInvalidTypeNotArray
-	 *
-	 * @covers \WpOrg\Requests\Session::request_multiple
-	 *
-	 * @param mixed $input Invalid parameter input.
-	 *
-	 * @return void
-	 */
-	public function testRequestMultipleInvalidOptions($input) {
-		$this->expectException(InvalidArgument::class);
-		$this->expectExceptionMessage('Argument #2 ($options) must be of type array');
-
-		$session = new Session();
-		$session->request_multiple([], $input);
-	}
-
-	/**
-	 * Data Provider.
-	 *
-	 * @return array
-	 */
-	public function dataInvalidTypeNotArray() {
-		return TypeProviderHelper::getAllExcept(TypeProviderHelper::GROUP_ARRAY);
 	}
 }
